In cumulus, integrate CoreToChip inside Chip.
Formerly, CoreToChip was seen as building the top-level netlist of the
chip only. But, when using special spacer pads, we need to have access
to that specific information from inside the Chip pads builder.
So we now move CoreToChip as an internal part of Chip. The right
CoreToChip to use (techno dependant) must be given as a configuration
parameter:

  "conf.coreToChipClass"

It is the class, not an object that must be suplied.