Add a prefix to GDS sub-cell isolated in a "per cell" library.
The GDS file write all the cell names (SREF) in one namespace, so cells
with identical names, coming from different libraries may clash. When
a GDS is read *and* contains sub-cell, we isolate them in a sub-library,
but we must *also* give the cells in them unique names, we prefix by
the library name (which in turn is the top cell name).