Merge the Anabatic state and Katana Stage into a stage.
The Anabatic::_state and Katana::_stage are partly redundant and confusive,
regroup them under Anabatic::_stage. All states of both engines are
merged. This allows to *not* update the "same ripup" while the router
is not running. This in turn render the propagateCagedConstraints() to
be less dependent from the initial positions.