Prevent *very* small nets to be broken in any way.
Two new configuration parameters define the size below which a net is
considered *very* small, and hence, none of its segments should be
broken. They would only lead to over-fragmentation. The "size" of a
net is the area enclosing all its RoutingPads.
    anabatic.smallNetWidth   (default: 0, disabled)
    anabatic.smallNetHeight  (default: 0, disabled)
Their value should be a fraction of the slice height.