Add support for variable metal distance around wide segments.
Implemented using a new kind of track marker : TrackMarkerSpacing,
which are kept in a separate vector. When Track::addOverlapCost()
is called it will then see if the parallel run length exceed the
threshold over those markers and flag the track cost as infinite.
It also inpact the perpandicular free computation. We don't want
the moving of a segment extend the perpandiculars length over the
threshold. This is done through Track::expandFreeInterval(),
restrict is so that the parallel length abide the rule. This may
need some more refinement here.