Fix drag constraints for "same layer VIAs/segments".
When an AutoContact can be dragged, that is vertical anchored on
vertical (likewise for horizonal), if the segment is of the same
layer as the anchor, we can ignore the VIAs end cap constraints
as it is enclosed in a wider area. This overconstraint was
preventing to achieve routing on I/O pins in rare corner cases.
Add an AutoSegment::SegDragSameLayer flag.